Another fix for initial statusbar
see http://vaniraosp.goo.im/#/c/127/ for description of issue

This is another (untested) fix, and it should keep the same choice layout as before but simply make don't show default.

Change-Id: Id955f4348acf4232810ad3ca27b5793748bef97e
NOTE: This and the other commit (127) is a one or the other thing, not both. So either choose this one for the same layout or the other for a different layout