if this is a feature we even want it needs to be pushed to gerrit for review by the team
merging source in settings and frameworks base is a bad idea.  these are our repos.
if the team wanted to build CM we could just build CM and not have to deal with all the bugs
this is introducting with little to no benefit at this point.  Also, it bypasses review and
makes our shit look sloppy as fuck along with creating a bunch of extra work.

this commit reverts "Add WhisperPush category to security settings. It's introducing dead code.

This reverts commit df8c02d0cb697f97b2cb7e61254d08a0f57ec97f.

Change-Id: Icbd7de303ba13043139040860b16cc84cc5f70b7