remove duplicate settings from merge
this hides the new respective security settings when N/A to the current
security method chosen.  It also removes duplicate settings and CM's
unnecessary isCmSecurity boolean.

it all works except when adding code to make vibrate work with PIN security.
this did not work with CM's code either even after the code to remove the
preference was disabled.

Change-Id: I7a5c56b72e2fc4ab08de52de88e9613ef6b9183f