[hack] temp fix for lockscreen blur animation
when blur is enabled and lockscreen transparency is disabled.. the wallpaper
flashes if you quickly turn the screen on/off when in an app.  it also flashes again
when the keyguard exits to an app.  it looks shitty.  the animated blur effect is also
broken under these conditions.  if you enable LOCKSCREEN_SEE_THROUGH it is fixed.

the frameworks code needs to be evaluated.. but it's 6AM and i haven't been able to sleep
yet so i'm going to give it another shot later today.

Change-Id: I59fe864977834a7d92c9bfa90f440683ac4d7a5d