logging toggling
toggle for logcat and dmsg to reduce overhead since it's not really
necessary for most people to keep these on all the time. uses
bootreceiver in settings to restore so if you FUBAR in your pants
you can still get a healthy log output up until the boot is completed.

PS2: moved to development settings and use shared preference vs system
setting
PS3: remove sharedpreference usage
PS4: remove the preference if the kernel doesn't support it

Change-Id: I82538b8be6e9dd99300fb2f9dd4c2b26a44c3551