fix some issues with the new navbar resizing class
my original resizing preference + fixes/improvements was redone to work better with hardkey
devices but some issues have been introduced. this commit fixes the navbar reinflating every
time you enter the fragment and also allows a smaller size on devices that a navbar is
required and not optional.

some of our github history was selectively deleted so the link to the original with
the fixes/improvements is here. http://vaniraosp.goo.im/#/c/10550/2

Change-Id: I78101e96c0cdb7385ff0368d0a36222f37dc4a65