settings: Unbreak wifi display on tablets
In change 46af8086db0db5b41f20c9f05a5f42c182262576 the option
to enable/disable the wireless display was moved to an action
bar which ends up entirely removing the ability to enable it
for devices using Settings in tablet (multi-pane) mode.

Follow the pattern of dreamcast which correctly displays
and removes the action bar.

Change-Id: Ie82b04b5e96a59ec093323601f8e79d23a6e69de

Conflicts:

	src/com/android/settings/wfd/WifiDisplaySettings.java