Gesture panel: setEnabled(false) when no gesture panel is added
There's no way to add gesture actions when we've added a panel to the various spots it
can be in. Lets setEnabled(false) until a gesture panel has been added. This lets the
user see where the gestures can be added but doesn't create confusion since the panels
can't be added from that fragment.

Change-Id: I80ac2bb5c9f3bc9fc3a967f84249e9e8fcd3b28b