Settings: set proper default value for tap to wake
The first time the display_settings.xml file gets inflated, the
PreferenceManager takes the defaultValue and persists it in the default
shared preferences file.

When we read the double tap to wake preference,
we assume it has a default value of true, which it uses until the user
enters Display Settings for the first time. At that point, the default
value is persisted, and every boot after will use the default value in
display_settings.xml.

Remove the default defined in the XML file.

Change-Id: I0989dd73b6b3c42ff8649bc17eebf4e191293ee4
Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:

	res/xml/display_settings.xml
	src/com/android/settings/DisplaySettings.java

Vanirification: revise mutual exclusion with proximityCheckOnWake