Fix NPE if Activity is chosen as action for LockscreenShortcut, but it does not have its own drawable since it is just an Activity. The app that it belongs to will show as the drawable now, instead of crashing.
But if someone chooses something like Browser.AddBookmarkPage and then
they go and swipe their new lockscreen shortcut to open the Activity,
it will of course force close because they are total, complete idiots!

But at least the menu choices no longer crash from choosing an Activity
that does not have it's own drawable and/or cannot be run on it's own.
Now you have to leave Security Settings, Settings, Activate LockScreen,
and then swipe it... so you are trying to be dumb if you go that far.

PS2: Shorter. Simpler. Works. Most activites work now. Later I can go
through them and fix or remove ones that don't.

Change-Id: Ie710137f7fb692bbb5d5c788cd11b9502da361c7
Signed-off-by: sudosurootdev <sudosurootdev@gmail.com>